+++ vault_cli/cli.py
@@ -16,6 +16,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
+import logging
 import os
 from typing import Any, Dict, Mapping, NoReturn, Sequence
 
@@ -23,6 +24,8 @@ import click
 import yaml
 
 from vault_cli import client, settings, types
+
+logger = logging.getLogger(__name__)
 
 CONTEXT_SETTINGS = {
     "help_option_names": ["-h", "--help"],
@@ -42,6 +45,13 @@ def load_config(ctx: click.Context, param: click.Parameter, value: str) -> None:
 
     config = settings.build_config_from_files(*config_files)
     ctx.default_map = config
+
+
+def set_verbosity(ctx: click.Context, param: click.Parameter, value: int) -> int:
+    level = settings.get_log_level(verbosity=value)
+    logging.basicConfig(level=level)
+    logger.info(f"Log level set to {logging.getLevelName(level)}")
+    return value
 
 
 @click.group(context_settings=CONTEXT_SETTINGS)
@@ -89,6 +99,14 @@ def load_config(ctx: click.Context, param: click.Parameter, value: str) -> None:
     help="Name of the backend to use (requests, hvac)",
 )
 @click.option(
+    "-v",
+    "--verbose",
+    is_eager=True,
+    callback=set_verbosity,
+    count=True,
+    help="Use multiple times to increase verbosity",
+)
+@click.option(
     "--config-file",
     is_eager=True,
     callback=load_config,
@@ -105,14 +123,18 @@ def cli(ctx: click.Context, **kwargs) -> None:
 
     """
     kwargs.pop("config_file")
+    verbose = kwargs.pop("verbose")
     backend: str = kwargs.pop("backend")
 
     kwargs.update(extract_special_args(ctx.default_map, os.environ))
 
     # There might still be files to read, so let's do it now
     kwargs = settings.read_all_files(kwargs)
+    saved_settings = kwargs.copy()
+    saved_settings.update({"backend": backend, "verbose": verbose})
     try:
         ctx.obj = client.get_client_from_kwargs(backend=backend, **kwargs)
+        ctx.obj.saved_settings = saved_settings
     except ValueError as exc:
         raise click.UsageError(str(exc))
 
@@ -277,6 +299,21 @@ def write_yaml(filepath, content) -> None:
 
 def exec_command(command: Sequence[str], environ: Dict[str, str]) -> NoReturn:
     os.execvpe(command[0], tuple(command), environ)
+
+
+@cli.command("dump-config")
+@click.pass_obj
+def dump_config(client_obj: client.VaultClientBase,) -> None:
+    """
+    Displays all the current settings in the format of a config file.
+    """
+    assert client_obj.saved_settings
+    click.echo(
+        yaml.safe_dump(
+            client_obj.saved_settings, default_flow_style=False, explicit_start=True
+        ),
+        nl=False,
+    )
 
 
 def main():